pyproject.toml
[project]
name = "tsp-dictionarylearning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cvxpy>=1.6.0",
    "einops>=0.8.1",
    "gudhi>=3.10.1",
    "gurobipy>=12.0.1",
    "h5py>=3.12.1",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.0",
    "networkx>=3.4.2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "seaborn>=0.13.2",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
